refactor: 로그인, 로그아웃, 회원가입 기능을 애노테이션 기반으로 리팩터링
