+++ app/src/main/java/com/techcourse/controller/UserController.java
@@ -1,0 +1,59 @@
+package com.techcourse.controller;
+
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.view.JspView;
+import com.techcourse.domain.User;
+import com.techcourse.repository.InMemoryUserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+@Controller
+public class UserController {
+
+    private static final Logger log = LoggerFactory.getLogger(UserController.class);
+
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ModelAndView login(final HttpServletRequest request, final HttpServletResponse response) {
+        if (UserSession.isLoggedIn(request.getSession())) {
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
+        }
+
+        return InMemoryUserRepository.findByAccount(request.getParameter("account"))
+                .map(user -> {
+                    log.info("User : {}", user);
+                    return login(request, user);
+                })
+                .orElse(new ModelAndView(new JspView("redirect:/401.jsp")));
+    }
+
+    private ModelAndView login(final HttpServletRequest request, final User user) {
+        if (user.checkPassword(request.getParameter("password"))) {
+            final var session = request.getSession();
+            session.setAttribute(UserSession.SESSION_KEY, user);
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
+        }
+        return new ModelAndView(new JspView("redirect:/401.jsp"));
+    }
+
+    @RequestMapping(value = "/logout", method = RequestMethod.POST)
+    public ModelAndView logout(final HttpServletRequest request, final HttpServletResponse response) {
+        final var session = request.getSession();
+        session.removeAttribute(UserSession.SESSION_KEY);
+        return new ModelAndView(new JspView("redirect:/index.jsp"));
+    }
+
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    public ModelAndView register(final HttpServletRequest request, final HttpServletResponse response) {
+        final var user = new User(2,
+                request.getParameter("account"),
+                request.getParameter("password"),
+                request.getParameter("email"));
+        InMemoryUserRepository.save(user);
+        return new ModelAndView(new JspView("redirect:/index.jsp"));
+    }
+}
